requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
greenlet==1.1.2
idna==3.3
psycopg2==2.9.3
python-decouple==3.6
requests==2.27.1
SQLAlchemy==1.4.32
urllib3==1.26.8
vk-api==11.9.7
